backend/requirements.txt
# Core FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validation & Serialization
pydantic[email]==2.5.0

# Environment Configuration
python-dotenv==1.0.0

# Authentication & Security
pyjwt==2.8.0
cryptography>=41.0.0
python-dateutil==2.8.2

# HTTP Client
httpx>=0.25.2

# Redis (for future sessions)
redis[hiredis]>=4.5.2,<5.0.0

# AI & Agent Framework
openai>=1.6.1
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langsmith>=0.1.0

# Structured Output & Parsing
instructor>=1.3.0
pydantic-ai>=0.0.12
